Split into public and private controllers
