+++ app/controllers/public_area_controller.rb
@@ -1,0 +1,2 @@
+class PublicAreaController < ApplicationController
+end
